// modules/configurator/models/CarModel.ts
export interface Car {
    id: string;
    name: string;
    year: number;
    image: string;
    defaultColor: string;   
    defaultColorFull: string;  
    defaultWheels: string;   
    defaultWheelsFull: string;   
    defaultInterior: string;
    defaultInteriorFull: string;
    carType: string; 
    price: number;
    colorPrice: number;
    wheelsPrice: number;
    interiorPrice: number;
  }